requirements.txt
openmeteo-requests
requests-cache
retry-requests
pandas
pyarrow
google-cloud-storage